+++ Reddit/Easy/making_links/src/lib.rs
@@ -1,6 +1,14 @@
 
 pub fn convert(input: &str) -> String {
-    println!("input: {}", input);
+    let mut input = input.parse::<i32>().unwrap();
+
+    while input > 0 {
+        println!("{}", input%62);
+
+        input /= 62;
+    }
+
+    // 0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ
 
     "".to_string()
 }