+++ Reddit/Easy/making_links/src/main.rs
@@ -1,5 +1,10 @@
 extern crate making_links;
 
+use std::io::{self, BufRead};
+
 fn main() {
-    println!("{}", making_links::convert("7026425611433322325"));
+    let stdin = io::stdin();
+    let line: String = stdin.lock().lines().map(|l| l.unwrap()).nth(0).unwrap();
+
+    println!("{}", making_links::convert(&line));
 }